move fixture back to test_catalog
