+++ tests/integ/test_catalog_rest_mode.py
@@ -14,7 +14,7 @@ from snowflake.snowpark import context
 from snowflake.snowpark._internal.analyzer.analyzer_utils import unquote_if_quoted
 from snowflake.snowpark.catalog import Catalog
 from snowflake.snowpark.types import IntegerType
-from tests.integ.catalog_integ_common import (
+from tests.integ.test_catalog import (
     CATALOG_TEMP_OBJECT_PREFIX,
     create_temp_db,
     create_temp_schema,
